feat(classify): replace sheet with dedicated new-run page
NewClassificationRunPage at /classify/new provides a full-page scrollable
form with ParseMethodSelector and ClassificationConfig. Navigation replaces
the 60vw sheet in both ClassificationPage (New Run) and
ClassificationRunDetailPage (Re-run with pre-populated defaults via location
state). ClassificationRunSheet deleted. Document preview deferred.
